+++ bin/create-github-release.mjs
@@ -1,14 +1,14 @@
 #!/usr/bin/env node
 let flux_shutdown_handler = null;
 try {
-    flux_shutdown_handler = (await import("flux-shutdown-handler/src/FluxShutdownHandler.mjs")).FluxShutdownHandler.new();
+    flux_shutdown_handler = await (await import("flux-shutdown-handler/src/FluxShutdownHandler.mjs")).FluxShutdownHandler.new();
 
     const path = process.argv[2] ?? null;
     if (path === null) {
         throw new Error("Please pass a path");
     }
 
-    await (await import("../src/CreateGithubRelease.mjs")).CreateGithubRelease.new()
+    await (await (await import("../src/CreateGithubRelease.mjs")).CreateGithubRelease.new())
         .createGithubRelease(
             path
         );