+++ bin/get-release-changelog.mjs
@@ -1,14 +1,14 @@
 #!/usr/bin/env node
 let flux_shutdown_handler = null;
 try {
-    flux_shutdown_handler = (await import("flux-shutdown-handler/src/FluxShutdownHandler.mjs")).FluxShutdownHandler.new();
+    flux_shutdown_handler = await (await import("flux-shutdown-handler/src/FluxShutdownHandler.mjs")).FluxShutdownHandler.new();
 
     const path = process.argv[2] ?? null;
     if (path === null) {
         throw new Error("Please pass a path");
     }
 
-    process.stdout.write(await (await import("../src/GetReleaseChangelog.mjs")).GetReleaseChangelog.new()
+    process.stdout.write(await (await (await import("../src/GetReleaseChangelog.mjs")).GetReleaseChangelog.new())
         .getReleaseChangelog(
             path
         ));