+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes:
 - Remove `bin` folder from build
 - Remove `node` user (Always run with host user id)
 - Merge `RUN`'s in `Dockerfile`
+- Make static new async
 
 ## v2024-01-15-1